Only allow single PurchaseOrderAPI per environment (#30)
* Add function to get all microservice deployments.

And also use it in the function to get a specific one.

* Add check whether there already is a POapi in environment.

Disregarding the name and microservice id. There should really only be
one of them right now.

* Don't need all those arguments

* We can also check the GitRepo for existing POapi.

It is a bit exessive to do both, but the first one should be faster and
if we don't check with kubernetes. We could get into weird states.

* Make sure I never forget a label here again

* Allow explicit no-release tag

* Restructure specs, more cases and att for new method.

* Add some logs in the PurchaseOrderAPI handler